future weather viewmodel

+++ app/src/main/java/com/dvt/forecastmvvm/data/db/entity/Day.kt
@@ -9,9 +9,9 @@ data class Day(
     @SerializedName("avgtemp_f")
     val avgtempF: Double,
     @SerializedName("avgvis_km")
-    val avgvisKm: Int,
+    val avgvisKm: Double,
     @SerializedName("avgvis_miles")
-    val avgvisMiles: Int,
+    val avgvisMiles: Double,
     @Embedded(prefix = "condition_")
     val condition: Condition,
     @SerializedName("maxtemp_c")
@@ -27,8 +27,8 @@ data class Day(
     @SerializedName("mintemp_f")
     val mintempF: Double,
     @SerializedName("totalprecip_in")
-    val totalprecipIn: Int,
+    val totalprecipIn: Double,
     @SerializedName("totalprecip_mm")
-    val totalprecipMm: Int,
+    val totalprecipMm: Double,
     val uv: Double
 )
